Ensuring writing to socket sends all of the data.
The write method was incorrectly implemeneted like a socket.write
instead of a file.write where there was a return value and not all data
might be written.  This was fixed to ensure all data would be written
with no return value to avoid characters being dropped in telnet
replies.